refactor(evaluatios): reorganize evaluation components and imports
