+++ components/evaluations/evaluation-list-item/evaluations.tsx
@@ -1,6 +1,6 @@
 import { Ban, Check } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 function EvaluationSymbols({ value }: { value: number }) {
   return (
