+++ components/evaluations/evaluation-list-item/tags.tsx
@@ -1,4 +1,4 @@
-import { Badge } from "../ui/badge";
+import { Badge } from "@/components/ui/badge";
 import React from "react";
 
 export default function Tags({
